Resolves Cordova Android issue.
Cordova Android (v4.2.2) does not have a css 'load' event, or it does and it is broken -- not sure. Mapbox can operate just fine without requiring the css to be loaded first. In fact this fix may cause the maps to load faster as the css will be done async.